+++ lib/atomic_web/templates/layout/_user_menu.html.heex
@@ -1,6 +1,5 @@
 <%= if @current_user do %>
   <li><%= @current_user.email %></li>
-  <li><%= link("Settings", to: Routes.user_settings_path(@conn, :edit)) %></li>
   <li><%= link("Log out", to: Routes.user_session_path(@conn, :delete), method: :delete) %></li>
 <% else %>
   <li><%= link("Register", to: Routes.user_registration_path(@conn, :new)) %></li>
