+++ lib/atomic_web/templates/user_change_password/edit.html.heex
@@ -1,0 +1,73 @@
+<div class="flex flex-col items-center w-screen h-screen justify-center bg-hero-pattern bg-cover lg:bg-[length:2000px_1400px] bg-no-repeat	bg-opacity-25 bg-center">
+  <.form let={f} for={@changeset} action={Routes.user_change_password_path(@conn, :update)} id="update_password" class="flex flex-col items-center justify-center w-full h-full mx-8 mt-8 space-y-6 sm:mx-0">
+    <h2 class="mx-4 mt-6 text-center text-5xl font-bold text-zinc-900 select-none">
+      <span><%= gettext("Change Password") %></span>
+    </h2>
+    <%= if @error_message do %>
+      <div class="w-full sm:px-0 px-4 sm:w-96 mx-4 mt-4 ">
+        <div class="flex flex-row w-full alert alert-danger text-sm bg-red-100 border-l-4 border-red-400 text-red-700 px-4 py-4">
+          <Heroicons.Solid.exclamation class="w-5 h-5 mr-2 text-red-400" />
+          <div>
+            <p><%= @error_message %></p>
+          </div>
+        </div>
+      </div>
+    <% end %>
+
+    <%= hidden_input(f, :action, value: "update_password") %>
+
+    <div class="space-y-3 rounded-md w-full flex-col mt-6">
+      <div class="flex flex-col items-center justify-center mx-4">
+        <div x-data="{ hide: true }" class="sm:w-96 flex flex-row items-center justify-center w-full border border-zinc-300 rounded placeholder-zinc-500 focus-within:z-10 focus-within:border-orange-400 focus-within:ring-orange-400 focus-within:ring-1 sm:text-sm">
+          <%= label(f, :current_password, "Current password", class: "sr-only") %>
+          <%= password_input(f, :current_password, required: true, placeholder: gettext("Current password"), class: "relative w-full border-0 text-zinc-900 rounded appearance-none outline-none focus:ring-0 sm:text-sm") %>
+          <div x-show="hide" x-on:click="hide = false; document.getElementById('update_password_current_password').type = 'text';" class="px-2 w-fit cursor-pointer">
+            <Ionicons.Filled.eye class="w-5 h-auto fill-black hover:fill-black" />
+          </div>
+          <div x-show="!hide" x-on:click="hide = true; document.getElementById('update_password_current_password').type = 'password';" class="px-2 w-fit cursor-pointer">
+            <Ionicons.Filled.eye_off class="w-5 h-auto fill-black hover:fill-black" />
+          </div>
+        </div>
+        <span class="text-sm w-full sm:w-96 mx-4 text-red-600"><%= error_tag(f, :current_password) %></span>
+      </div>
+
+      <div class="flex flex-col items-center justify-center mx-4">
+        <div x-data="{ hide: true }" class="sm:w-96 flex flex-row items-center justify-center w-full border border-zinc-300 rounded placeholder-zinc-500 focus-within:z-10 focus-within:border-orange-400 focus-within:ring-orange-400 focus-within:ring-1 sm:text-sm">
+          <%= label(f, :password, "New password", class: "sr-only") %>
+          <%= password_input(f, :password, required: true, placeholder: gettext("New password"), class: "relative w-full border-0 text-zinc-900 rounded appearance-none outline-none focus:ring-0 sm:text-sm") %>
+          <div x-show="hide" x-on:click="hide = false; document.getElementById('update_password_password').type = 'text';" class="px-2 w-fit cursor-pointer">
+            <Ionicons.Filled.eye class="w-5 h-auto fill-black hover:fill-black" />
+          </div>
+          <div x-show="!hide" x-on:click="hide = true; document.getElementById('update_password_password').type = 'password';" class="px-2 w-fit cursor-pointer">
+            <Ionicons.Filled.eye_off class="w-5 h-auto fill-black hover:fill-black" />
+          </div>
+        </div>
+        <span class="text-sm w-full sm:w-96 mx-4 text-red-600"><%= error_tag(f, :password) %></span>
+      </div>
+
+      <div class="flex flex-col items-center justify-center mx-4">
+        <div x-data="{ hide_confirm: true }" class="sm:w-96 flex flex-row items-center justify-center w-full border border-zinc-300 rounded placeholder-zinc-500 focus-within:z-10 focus-within:border-orange-400 focus-within:ring-orange-400 focus-within:ring-1 sm:text-sm">
+          <%= label(f, :password_confirmation, "Confirm new password", class: "sr-only") %>
+          <%= password_input(f, :password_confirmation,
+            required: true,
+            placeholder: gettext("Confirm new password"),
+            class: "relative w-full border-0 text-zinc-900 rounded appearance-none outline-none focus:ring-0 sm:text-sm"
+          ) %>
+          <div x-show="hide_confirm" x-on:click="hide_confirm = false; document.getElementById('update_password_password_confirmation').type = 'text';" class="px-2 w-fit cursor-pointer">
+            <Ionicons.Filled.eye class="w-5 h-auto fill-black hover:fill-black" />
+          </div>
+          <div x-show="!hide_confirm" x-on:click="hide_confirm = true; document.getElementById('update_password_password_confirmation').type = 'password';" class="px-2 w-fit cursor-pointer">
+            <Ionicons.Filled.eye_off class="w-5 h-auto fill-black hover:fill-black" />
+          </div>
+        </div>
+        <span class="text-sm w-full sm:w-96 mx-4 text-red-600"><%= error_tag(f, :password_confirmation) %></span>
+      </div>
+    </div>
+
+    <div class="w-full flex-col">
+      <div class="mx-4 justify-center flex select-none">
+        <%= submit("Change Password", class: "w-full flex justify-center border-2 rounded-md bg-orange-500 text-lg border-orange-500 py-2 px-3.5 text-sm font-medium text-white shadow-sm sm:w-96") %>
+      </div>
+    </div>
+  </.form>
+</div>
