+++ lib/atomic_web/views/user_change_password_view.ex
@@ -1,0 +1,3 @@
+defmodule AtomicWeb.UserChangePasswordView do
+  use AtomicWeb, :view
+end
