--- lib/atomic_web/views/user_settings_view.ex
@@ -1,3 +1,0 @@
-defmodule AtomicWeb.UserSettingsView do
-  use AtomicWeb, :view
-end
